finalized intersect/sliced function
