+++ requirements.txt
@@ -5,3 +5,4 @@ numpy
 pyproj
 netCDF4
 xarray
+shapely
